chore: 一部のテストのタイムアウトを長くする
